Move SchemaUnawarePostGISDialect and PostgisDataStoreProvider as top level classes
Being both PostgisBackendAutoConfiguration inner classes makes reading
difficult
